feat(AdminUserPanel): add toast notification
