+++ apps/web/package.json
@@ -26,6 +26,7 @@
 		"formik": "^2.2.9",
 		"rc-progress": "^3.4.1",
 		"react-spinners": "^0.13.8",
+		"react-toastify": "^9.1.1",
 		"yup": "^0.32.11"
 	},
 	"devDependencies": {
